feat(start-dialog): add a separate component for the start dialog
